pyproject.toml
[project]
name = "travel-agent"
version = "0.1.0"
description = "LangGraph multi-agent travel planner (supervisor + sub-agents)"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langchain-core>=0.3",
    "google-search-results>=2.4.2",
    "pydantic>=2",
    "python-dotenv>=1.0",
    "gradio>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/travel_agent"]
